+++ wannago-server/src/main/java/com/wannago/qna/answer/service/AnswerServiceImpl.java
@@ -3,13 +3,13 @@ package com.wannago.qna.answer.service;
 import com.wannago.member.entity.Member;
 import com.wannago.qna.answer.dto.AnswerRequest;
 import com.wannago.qna.answer.repository.AnswerRepository;
-import com.wannago.qna.question.repository.AskRepository;
 import com.wannago.member.repository.MemberRepository;
 import com.wannago.common.exception.CustomErrorCode;
 import com.wannago.common.exception.CustomException;
 import com.wannago.qna.answer.dto.AnswerResponse;
-import com.wannago.qna.entity.Answer;
-import com.wannago.qna.entity.Ask;
+import com.wannago.qna.ask.entity.Answer;
+import com.wannago.qna.ask.entity.Ask;
+import com.wannago.qna.ask.repository.AskRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
