+++ wannago-server/src/main/java/com/wannago/qna/ask/controller/AskController.java
@@ -1,10 +1,11 @@
-package com.wannago.qna.controller;
+package com.wannago.qna.ask.controller;
 
-import com.wannago.dto.AskRequestDto;
-import com.wannago.dto.AskResponseDto;
-import com.wannago.service.AskService;
+
+import com.wannago.qna.ask.dto.AskRequest;
+import com.wannago.qna.ask.dto.AskResponse;
+import com.wannago.qna.ask.service.AskService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity; // ResponseEntity를 import 합니다.
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -18,14 +19,13 @@ public class AskController {
 
     //질문 등록
     @PostMapping
-    public ResponseEntity<AskResponseDto> createAsk(@RequestBody AskRequestDto requestDto) {
-        // askService에서 생성된 DTO를 '쟁반'에 담아 성공(OK) 상태와 함께 반환합니다.
+    public ResponseEntity<AskResponse> createAsk(@RequestBody AskRequest requestDto) {
         return ResponseEntity.ok(askService.createAsk(requestDto));
     }
 
     //질문 수정
     @PutMapping("/{id}")
-    public ResponseEntity<AskResponseDto> updateAsk(@PathVariable Long id, @RequestBody AskRequestDto requestDto) {
+    public ResponseEntity<AskResponse> updateAsk(@PathVariable Long id, @RequestBody AskRequest requestDto) {
         return ResponseEntity.ok(askService.updateAsk(id, requestDto));
     }
 
@@ -38,13 +38,13 @@ public class AskController {
 
     //질문 목록 조회
     @GetMapping
-    public ResponseEntity<List<AskResponseDto>> getAsks() {
+    public ResponseEntity<List<AskResponse>> getAsks() {
         return ResponseEntity.ok(askService.getAsks());
     }
 
     //질문 상세 조회
     @GetMapping("/{id}")
-    public ResponseEntity<AskResponseDto> getAsk(@PathVariable Long id) {
+    public ResponseEntity<AskResponse> getAsk(@PathVariable Long id) {
         return ResponseEntity.ok(askService.getAsk(id));
     }
 }
