+++ wannago-server/src/main/java/com/wannago/qna/ask/dto/AskRequest.java
@@ -1,0 +1,14 @@
+package com.wannago.qna.ask.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class AskRequest {
+    private String category;
+    private String title;
+    private String content;
+}
