+++ wannago-server/src/main/java/com/wannago/qna/ask/entity/Ask.java
@@ -1,9 +1,7 @@
-package com.wannago.qna.entity;
+package com.wannago.qna.ask.entity;
 
-import com.wannago.member.entity.Member;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
@@ -15,26 +13,24 @@ import java.time.LocalDateTime;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
-@Entity(name = "answer")
+@Entity(name = "ask")
 @EntityListeners(AuditingEntityListener.class)
-public class Answer {
+public class Ask {
 
     @Id
-    @Column(name = "answer_id")
+    @Column(name = "ask_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "ask_id", nullable = false)
-    private Ask askId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member memberId;
+    @Enumerated(EnumType.STRING)
+    @Column(name = "category", nullable = false)
+    private Category category;
 
     @Column(name = "author", length = 30, nullable = false)
     private String author;
+
+    @Column(name = "title", length = 100, nullable = false)
+    private String title;
 
     @Column(name = "contents", columnDefinition = "TEXT", nullable = false)
     private String contents;
@@ -50,13 +46,15 @@ public class Answer {
     @Column(name = "modified_date", columnDefinition = "TIMESTAMP", nullable = false)
     private LocalDateTime modifiedDate;
 
-    // 답변 수정을 위한 메서드
-    public void updateContents(String contents) {
+    // 필드 값 변경을 위한 update 메서드 추가
+    public void update(Category category, String title, String contents) {
+        this.category = category;
+        this.title = title;
         this.contents = contents;
     }
-
-    // 채택 처리를 위한 메서드
-    public void accept() {
-        this.isAccepted = true;
+    public Ask(Category category, String title, String contents) {
+        this.category = category;
+        this.title = title;
+        this.contents = contents;
     }
 }
