--- wannago-server/src/main/java/com/wannago/qna/entity/Category.java
@@ -1,5 +1,0 @@
-package com.wannago.qna.entity;
-
-public enum Category {
-    ACTIVITY, FOOD, LOCATION, OTHER;
-}
