--- wannago-server/src/main/java/com/wannago/qna/question/repository/AskRepository.java
@@ -1,7 +1,0 @@
-package com.wannago.qna.question.repository;
-
-import com.wannago.qna.entity.Ask;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AskRepository extends JpaRepository<Ask,Long> {
-}
